refactor: User의 구현체 UserDomain으로 통일 및 status 검증 로직 도메인단으로 이전
