+++ src/main/java/com/newzet/api/user/business/service/UserFactory.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
-import com.newzet.api.user.domain.ActiveUser;
 import com.newzet.api.user.domain.User;
+import com.newzet.api.user.domain.UserDomain;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -15,6 +15,6 @@ import lombok.NoArgsConstructor;
 public class UserFactory {
 
 	public static User create(UUID id, String email, String nickname, String status) {
-		return ActiveUser.create(id, email, nickname);
+		return UserDomain.create(id, email, nickname, status);
 	}
 }
